pyproject.toml
[project]
name = "tramontane"
version = "0.1.0"
description = "Mistral-native agent orchestration framework"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = [
    "mistral", "agents", "orchestration", "llm",
    "gdpr", "eu-sovereign", "pipeline", "ai-framework",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mistralai>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "anyio>=4.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0.0"]
postgres = ["asyncpg>=0.29.0", "pgvector>=0.2.0"]
voice = ["sounddevice>=0.4.0", "numpy>=1.26.0"]
sandbox = ["e2b>=0.17.0"]
hub = ["huggingface-hub>=0.20.0"]

[project.scripts]
tramontane = "tramontane.cli.main:app"

[project.urls]
Homepage = "https://github.com/bleucommerce/tramontane"
Repository = "https://github.com/bleucommerce/tramontane"
Documentation = "https://github.com/bleucommerce/tramontane#readme"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "types-pyyaml>=6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
